switch file names back to not mess with lm policy
